requirements.txt
streamlit
pandas
numpy
opencv-python-headless
Pillow
streamlit-image-coordinates
streamlit-drawable-canvas
plotly